+++ app/controllers/web/users/recovery_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Web
+  module Users
+    class RecoveryController < Web::BaseController
+      skip_before_action :authenticate
+
+      def new; end
+    end
+  end
+end